marketing/ppc-manager/requirements.txt
fastmcp>=2.0.0
google-api-python-client>=2.140.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-ads>=25.1.0
google-analytics-data>=0.18.0
google-analytics-admin>=0.23.0
facebook_business>=21.0.0
cryptography>=43.0.0
filelock>=3.15.0
httpx>=0.27.0
python-dotenv>=1.0.0
